API

This page is a dump of all the docstrings found in the code. This will be cleaned up in the future and organized to be more accessible.

Mercury.LoopRateLimiter — Type

LoopRateLimiter

Runs a loop at a fixed rate. Works best for loops where the runtime is approximately the same every iteration. The loop runtime is kept approximately constant by sleeping for any time not used by the core computation. This is useful for situations where the computation should take place a regular, predictable intervals.

To achieve better accuracy, the rate limiter records the error between the expected runtime and actual runtime every N_batch iterations, and adjusts the sleep time by the average. Unlike the standard sleep function in Julia, this limiter has a minimum sleep time of 1 microsecond, and rates above 1000Hz can be achieved with moderate accuracy.

Example

lrl = LoopRateLimiter(100)  # Hz
reset!(lrl)
for i = 1:100
    startloop(lrl)          # start timing the loop
    myexpensivefunction()   # execute the core body of the loop
    sleep(lrl)              # sleep for the rest of the time
end

Mercury.Publisher — Type

Publisher

A simple wrapper around a ZMQ publisher, but only publishes protobuf messages.

Construction

Publisher(context::ZMQ.Context, ipaddr, port; name)

To create a publisher, pass in a ZMQ.Context, which allows all related publisher / subscribers to be collected in a "group." The publisher also needs to be provided the IPv4 address (either as a string or as a Sockets.IPv4 object), and the port (either as an integer or a string).

A name can also be optionally provided via the name keyword, which can be used to provide a helpful description about what the publisher is publishing. It defaults to "publisher_#" where # is an increasing index.

If the port

Usage

To publish a message, just use the publish method on a protobuf type:

publish(pub::Publisher, proto_msg::ProtoBuf.ProtoType)

Mercury.Subscriber — Type

Subscriber

A simple wrapper around a ZMQ subscriber, but only for protobuf messages.

Construction

Subscriber(context::ZMQ.Context, ipaddr, port; name)

To create a subscriber, pass in a ZMQ.Context, which allows all related publisher / subscribers to be collected in a "group." The subscriber also needs to be provided the IPv4 address (either as a string or as a Sockets.IPv4 object), and the port (either as an integer or a string).

A name can also be optionally provided via the name keyword, which can be used to provide a helpful description about what the subscriber is subscribing to. It defaults to "subscriber_#" where # is an increasing index.

Usage

Use the blocking subscribe method to continually listen to the socket and store data in a protobuf type:

subscribe(sub::Subscriber, proto_msg::ProtoBuf.ProtoType)

Note that this function contains an infinite while loop so will block the calling thread indefinately. It's usually best to assign the process to a separate thread / task:

sub_task = @task subscribe(sub, proto_msg)
schedule(sub_task)

Base.sleep — Method

sleep(::LoopRateLimiter)

Sleep the OS for the amount of time needed to achieve the rate specified by the loop rate limiter. Has a minimum sleep time of 1 microsecond (relies on the usleep C function).

Mercury.reset! — Method

reset!(::LoopRateLimiter)

Reset the loop rate limiter before a loop. Not necessary if the object is created directly before calling the loop.

Mercury.startloop — Method

startloop(::LoopRateLimiter)

Call this function at the beginning of a loop body to start timing the loop.

Mercury.usleep — Method

usleep(us)

Sleep for us microseconds. A wrapper around the C usleep function in unistd.h.

Mercury.@rate — Macro

@rate

Run a loop at a fixed rate, specified either by an integer literal or a LoopRateLimiter object. It will run the loop so that it executes close to rate iterations per second.

Examples

@rate for i = 1:100
    myexpensivefunction()
end 200#Hz

lr = LoopRateLimiter(200, N_batch=10)
@rate while i < 100
    myexpensivefunction()
    i += 1
end lr

Note that the following does NOT work:

rate = 100
@rate for i = 1:100
    myexpensivefunction()
end rate 

Since Julia macros dispatch on the compile-time types instead of the run-time types.